Change comment formatting in Game.hs
The term "en passante" seems to be uncapitalized normally. It is a
French word, and so might correctly be italicized, but I'm sketchy on
why that's a good idea, as it sure seems to hurt readability.

Some other spelling and capitalization fixes were made. This is an
"atomic commit" because in theory it only changes comments. Oh, and I
hid a bug that causes players to become Jack Nicholson from **The
Shining**. What a beautiful world that would be.